+++ lib/Model/InvoicePaymentTransaction.php
@@ -61,10 +61,10 @@ class InvoicePaymentTransaction implements ModelInterface, ArrayAccess
 'paymentId' => 'string',
 'paymentExternalKey' => 'string',
 'transactionType' => 'string',
-'amount' => 'BigDecimal',
+'amount' => 'float',
 'currency' => 'string',
 'effectiveDate' => '\DateTime',
-'processedAmount' => 'BigDecimal',
+'processedAmount' => 'float',
 'processedCurrency' => 'string',
 'status' => 'string',
 'gatewayErrorCode' => 'string',
@@ -1190,7 +1190,7 @@ self::STATUS_PAYMENT_SYSTEM_OFF,        ];
     /**
      * Gets amount
      *
-     * @return BigDecimal
+     * @return float
      */
     public function getAmount()
     {
@@ -1200,7 +1200,7 @@ self::STATUS_PAYMENT_SYSTEM_OFF,        ];
     /**
      * Sets amount
      *
-     * @param BigDecimal $amount Transaction amount, required except for void operations
+     * @param float $amount Transaction amount, required except for void operations
      *
      * @return $this
      */
@@ -1271,7 +1271,7 @@ self::STATUS_PAYMENT_SYSTEM_OFF,        ];
     /**
      * Gets processedAmount
      *
-     * @return BigDecimal
+     * @return float
      */
     public function getProcessedAmount()
     {
@@ -1281,7 +1281,7 @@ self::STATUS_PAYMENT_SYSTEM_OFF,        ];
     /**
      * Sets processedAmount
      *
-     * @param BigDecimal $processedAmount processedAmount
+     * @param float $processedAmount processedAmount
      *
      * @return $this
      */
